refactor(tests): move helpers and tests to deparate folders
