+++ src/shared/helpers/compareObjects/compareObjects.js
@@ -1,0 +1,43 @@
+export function compareObjects(obj1, obj2) {
+  if (obj1 === obj2) return true;
+
+  if (
+    typeof obj1 !== 'object' ||
+    typeof obj2 !== 'object' ||
+    typeof obj1 === 'undefined' ||
+    typeof obj2 === 'undefined'
+  ) {
+    return false;
+  }
+
+  const keysA = Object.keys(obj1);
+  const keysB = Object.keys(obj2);
+
+  if (keysA.length !== keysB.length) {
+    return false;
+  }
+
+  let result = true;
+
+  keysA.forEach((key) => {
+    if (!keysB.includes(key)) {
+      result = false;
+    }
+
+    if (isFunction(obj1[key]) || isFunction(obj2[key])) {
+      if (obj1[key].tostring() !== obj2[key].tostring()) {
+        result = false;
+      }
+    }
+
+    if (!compareObjects(obj1[key], obj2[key])) {
+      result = false;
+    }
+  });
+
+  return result;
+}
+
+function isFunction(value) {
+  return value === 'function';
+}
