+++ src/shared/helpers/getCurrentYear/getCurrentYear.js
@@ -1,0 +1,3 @@
+export function getCurrentYear() {
+  return new Date().getFullYear();
+}
